backend/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
websockets==12.0
pydantic==2.5.0
pydantic-settings==2.1.0

# AI Models
transformers==4.36.0
torch>=2.0.0
torchaudio>=2.0.0
torchvision>=0.15.0
accelerate==0.25.0
sentencepiece==0.1.99
protobuf==4.25.1

# Speech Processing
openai-whisper==20231117
TTS==0.22.0  # Coqui TTS
librosa==0.10.1
soundfile==0.12.1
pyaudio==0.2.14

# Image Processing
Pillow==10.1.0
opencv-python==4.8.1.78
scikit-learn==1.3.2

# LLM
llama-cpp-python==0.2.20

# Utilities
numpy==1.24.3
scipy==1.11.4
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
httpx==0.25.2
redis==5.0.1
celery==5.3.4
psutil==5.9.6
tqdm==4.66.1

# Language Processing
langdetect==1.0.9
mecab-python3==1.0.8
unidic-lite==1.0.8
googletrans==4.0.0-rc1

# Development
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1